Refactor monetary types to use decimal instead of double
- Updated namespace declarations for consistency.
- Changed Price properties in multiple classes from double to decimal for better precision.
- Modified GetTopByMaxPrice method signatures to accept decimal.
- Reformatted CreateCaseRequest class for clarity.
- Ensured consistent use of decimal in monetary calculations across services.
- Updated UpdateBalance method to accept decimal for amount.

+++ Interfaces/IItemRepository.cs
@@ -7,6 +7,6 @@ public interface IItemRepository
 {
     Task<CaseItem> Create(CaseItem item);
     Task<CaseItem?> GetById(ObjectId id);
-    Task<List<CaseItem>> GetTopByMaxPrice(double maxPrice, int limit);
+    Task<List<CaseItem>> GetTopByMaxPrice(decimal maxPrice, int limit);
     Task<List<CaseItem>> Get(int fromPrice = 0, int page = 1, int pageSize = 20);
 }
