+++ Interfaces/IUserService.cs
@@ -12,7 +12,7 @@ public interface IUserService
     Task<User> Create(User user);
     Task<User?> GetById(ObjectId id);
     Task<UserInfo> GetUserInfo(string userId);
-    Task<bool> UpdateBalance(ObjectId id, double amount);
+    Task<bool> UpdateBalance(ObjectId id, decimal amount);
     Task<List<InventoryItemView>> GetInventoryItems(string userId, int page = 1, int pageSize = 32);
     Task AddToInventory(ObjectId userId, List<CaseItemView> items);
     Task RemoveFromInventory(User user, ObjectId itemId, int quantity = 1);
