+++ Models/UserInfo.cs
@@ -6,7 +6,7 @@ namespace CaseBattleBackend.Models;
 public class UserInfo
 {
     public required string Id { get; set; }
-    public required double Balance { get; set; }
+    public required decimal Balance { get; set; }
     public required string Nickname { get; set; }
     public required Uri AvatarUrl { get; set; }
     public required int Level { get; set; }
